third_party/blink/web_tests/external/wpt/css/css-tables/tentative/table-fixed-minmax-expected.txt

This is a testharness.js-based test.
PASS table's width inside flexbox width:auto is not infinite
FAIL table's width inside flexbox width:max-content is infinite assert_greater_than: expected a number greater than 769 but got 100
PASS table's width inside flexbox width:min-content is infinite, but the flexbox is finite; scenario 1
PASS table's width inside flexbox width:min-content is infinite, but the flexbox is finite; scenario 2
PASS table's width inside grid width:auto is not infinite
PASS table's width inside grid width:max-content is infinite
PASS table's width inside a table cell is infinite
PASS table's width inside an absolute block is infinite
Harness: the test ran to completion.
